v.1.0.10 update
## Official Release v.1.0.10

***BREAKING (only if you have a saved session in progress)***
When the session is saved, new data is being stored in this update. Because the saved data has changed, sessions saved with the previous version of the app WILL NOT LOAD because they don't have this new data (the Continue button will be grayed out). So if you have a game in progress using a saved session, finish that mission before updating to this version of the app.

### *NEW BALANCING SYSTEM*
### Fame System: Adaptive Difficulty
---
Adaptive Difficulty is an interpretation of the official app’s "fame" system. The better the heroes get at eliminating Imperial groups, the more resources are deployed to stop them, but the better the reward. Adaptive Difficulty can be used with any difficulty setting. It defaults to OFF, but can be toggled ON during Mission Setup.

- Each time an enemy group is defeated, the Rebels' FAME increases.
- Each time an enemy group is defeated, Imperial Threat increases by a predetermined value.

Rebels gain AWARD points based on the FAME earned, which can be spent at the end of the Mission to buy Items and Side Missions.

### Balance Changes
-Cost changes, priority changes and Tier changes for various enemy groups and villains.
-Events now have a 1 in 4 chance to trigger instead of a 1 in 6 chance.
-The "Priority Shift" Event now has a higher chance to trigger than before (and it can trigger twice).
-The "Leave Them To Me" Event has been toned down. There is now an upper limit to the cost of the villain that it can deploy, based on the Imperial's current threat + 7.
-Because Vader's cost has been lowered in this update, the "As You Wish" Event further lowers his cost by 2 instead of 5.

### *Normal Updates*
**NEW**
- Ally cards can now be viewed just like deployed Imperial group cards - right click or swipe across the Ally's thumbnail.
- Clicking/tapping the VERSION info on the title screen (lower right corner) will now show you a CHANGE LOG for the latest version.

**CHANGES**
- The Save/Load mission defaults buttons are slightly larger to accommodate smaller screens.
- The zoomed card view shows the card slightly larger than before to accommodate smaller screens.

**FIXES**
- Events that have already triggered are now saved with the session state so they won't trigger again upon restoring the session (Exception: "Priority Shift" can trigger up to 2 times).
- The debug window wasn't closing when hitting SPACEBAR, and would instead keep populating the window with thumbnails